Update11

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -10,7 +10,7 @@
  */
 void puts2(char *str)
 {
-	unsigned long int i;
+	int i;
 
 	for (i = 0; str[i] != '\0'; i++)
 	{
